Switch rmsnorm weights to float
This doesn't have a tangible impact on precision, but this will make
future conversions of actual weights a little easier to do mechanically
as we don't want these to use fp8 et al.

This costs us ~20us/tok (out of ~16ms), which is something that may need
to be reconsidered later.

+++ tools/convert.py
@@ -139,17 +139,19 @@ dtype = {"bfloat16": torch.bfloat16, "float16": torch.float16, "float32": torch.
 tensors["model.embed.weight"] = weights["model.embed_tokens.weight"].to(dtype)
 
 for l in range(config["num_hidden_layers"]):
-    tensors[f"model.layers.{l}.attn.norm.weight"] = weights[f"model.layers.{l}.input_layernorm.weight"].to(dtype)
+    tensors[f"model.layers.{l}.attn.norm.weight"] = weights[f"model.layers.{l}.input_layernorm.weight"].float()
     tensors[f"model.layers.{l}.attn.wq.weight"] = permute_reverse(weights[f"model.layers.{l}.self_attn.q_proj.weight"], config["num_attention_heads"]).to(dtype)
     tensors[f"model.layers.{l}.attn.wk.weight"] = permute_reverse(weights[f"model.layers.{l}.self_attn.k_proj.weight"], config["num_key_value_heads"]).to(dtype)
     tensors[f"model.layers.{l}.attn.wv.weight"] = weights[f"model.layers.{l}.self_attn.v_proj.weight"].to(dtype)
     tensors[f"model.layers.{l}.attn.wo.weight"] = weights[f"model.layers.{l}.self_attn.o_proj.weight"].to(dtype)
-    tensors[f"model.layers.{l}.mlp.norm.weight"] = weights[f"model.layers.{l}.post_attention_layernorm.weight"].to(dtype)
+
+    tensors[f"model.layers.{l}.mlp.norm.weight"] = weights[f"model.layers.{l}.post_attention_layernorm.weight"].float()
+
     tensors[f"model.layers.{l}.mlp.w1.weight"] = weights[f"model.layers.{l}.mlp.gate_proj.weight"].to(dtype)
     tensors[f"model.layers.{l}.mlp.w2.weight"] = weights[f"model.layers.{l}.mlp.down_proj.weight"].to(dtype)
     tensors[f"model.layers.{l}.mlp.w3.weight"] = weights[f"model.layers.{l}.mlp.up_proj.weight"].to(dtype)
 
-tensors["model.norm.weight"] = weights["model.norm.weight"].to(dtype)
+tensors["model.norm.weight"] = weights["model.norm.weight"].float()
 tensors["model.output.weight"] = weights["lm_head.weight"].to(dtype)
 
 # metadata values must be strings in safetensors
